add onBook prop, remove mantine notification dependency
