+++ packages/react-scheduling/package.json
@@ -94,7 +94,6 @@
   "peerDependencies": {
     "@mantine/core": "^8.0.0",
     "@mantine/hooks": "^8.0.0",
-    "@mantine/notifications": "^8.0.0",
     "@medplum/core": "5.1.32",
     "@medplum/fhirtypes": "5.1.32",
     "@medplum/react": "5.1.32",
