+++ packages/react-scheduling/src/SchedulingWorkspace/SchedulingWorkspace.stories.tsx
@@ -1,6 +1,9 @@
 // SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
 // SPDX-License-Identifier: Apache-2.0
+import { showNotification } from '@mantine/notifications';
+import type { Appointment } from '@medplum/fhirtypes';
 import type { Meta } from '@storybook/react';
+import { IconCalendarCheck } from '@tabler/icons-react';
 import type { JSX } from 'react';
 import { withBookStub, withFindStub, withFixtures, withMockedDate } from '../stories/decorators';
 import { CalendarWeekFixtures, PatientFixtures, SchedulingFixtures } from '../stories/scheduling';
@@ -28,7 +31,9 @@ export default {
  * "Find a time": the search opens on that day rather than today, and the pane widens to
  * lay the times beside the form. Pick one, name a patient ("Jordan"), and book — the
  * pane closes and the appointment appears on the calendar without a reload, because the
- * booking announces what it wrote and the calendar is listening.
+ * booking announces what it wrote and the calendar is listening. The toast naming the
+ * visit is this story's, raised from `onBooked`: the workspace reports what was booked
+ * and leaves how to say so to whatever is hosting it.
  *
  * Clicking a different day with the form part-filled re-opens it on the new day and
  * clears the answers; clicking again inside the day already open leaves them alone.
@@ -41,6 +46,36 @@ export const Basic = (): JSX.Element => (
   // inside it, and a short host makes each of them look cramped for reasons of its own.
   // 72px is the package banner `withSchedulingHeader` puts above every story here.
   <div style={{ height: 'calc(100vh - 72px)', padding: '1em', boxSizing: 'border-box' }}>
-    <SchedulingWorkspace />
+    <SchedulingWorkspace
+      onBooked={({ appointment }) => {
+        showNotification({
+          color: 'green',
+          icon: <IconCalendarCheck size={18} />,
+          title: 'Appointment booked',
+          message: describeBooking(appointment),
+        });
+      }}
+    />
   </div>
 );
+
+/**
+ * Names the visit that was booked, for the notification announcing it.
+ * @param appointment - The appointment `$book` wrote.
+ * @returns Who it is for and when, as far as each is known.
+ */
+function describeBooking(appointment: Appointment): string {
+  const patient = (appointment.participant ?? []).find((participant) =>
+    participant.actor?.reference?.startsWith('Patient/')
+  );
+  const when = appointment.start
+    ? new Date(appointment.start).toLocaleString(undefined, {
+        weekday: 'long',
+        month: 'long',
+        day: 'numeric',
+        hour: 'numeric',
+        minute: '2-digit',
+      })
+    : undefined;
+  return [patient?.actor?.display, when].filter(Boolean).join(' · ');
+}
